Initial commit after separating /rest and /api endpoints including support for oas3 and swagger2 in each.
